Update ecs-url.md
This change is intended to make explicit the treatment of url.query when a compound query is present. 

Ambiguity exists in that a field mapped as keyword could be split into an array of kv-pairs and still maintain both the example text & the keyword mapping. 

By making plain the single string nature of the field contents, the reduction in ambiguity will help in resolving questions that may arise when formatting data to the schema.